+++ tsup.config.ts
@@ -6,6 +6,7 @@ export default defineConfig([
     entry: {
       index: 'src/index.ts',
       reporter: 'src/reporter.ts',
+      cli: 'src/cli.ts',
     },
     outDir: 'dist/esm',
     format: ['esm'],